Made it a stand-alone plugin. You can package it with Tree Commands.
Made it this way, as otherwise if two plugins have this package only one
will be loaded. This wouldn't be too bad, but some of the methods
(getBundle()) relied on the classLoader of the current class. This
means, one of the two plugins will stop working. This is now fixed by
passing your ClassLoader through the constructor.
It is ugly, I know, but I couldn't think of a better way.

Also added a "reload()" method for the file class bundle. This means you
don't need to create a new instance, but just call I18N#reload() to
apply any changes in the files.