TabCompleter: Added an option to perform a starting with check
If you set the boolean to true, it will only pick the tabcompletions
which start with the user's input. So if the completions are "dummy" and
"test" and the user writes "du" and tab-completes, only "dummy" will be
suggested. If he writes "te" and tab-completes, only "test" will be
suggested.

Default is true.